actions: Split out zerver.actions.realm_settings.
