+++ corporate/tests/test_stripe.py
@@ -86,8 +86,9 @@ from zerver.actions.create_user import (
     do_create_user,
     do_reactivate_user,
 )
+from zerver.actions.realm_settings import do_deactivate_realm, do_reactivate_realm
 from zerver.actions.users import do_deactivate_user
-from zerver.lib.actions import do_create_realm, do_deactivate_realm, do_reactivate_realm
+from zerver.lib.actions import do_create_realm
 from zerver.lib.test_classes import ZulipTestCase
 from zerver.lib.timestamp import datetime_to_timestamp, timestamp_to_datetime
 from zerver.lib.utils import assert_is_not_none
