+++ zerver/tests/test_audit_log.py
@@ -19,6 +19,14 @@ from zerver.actions.create_user import (
 )
 from zerver.actions.realm_icon import do_change_icon_source
 from zerver.actions.realm_playgrounds import do_add_realm_playground, do_remove_realm_playground
+from zerver.actions.realm_settings import (
+    do_deactivate_realm,
+    do_reactivate_realm,
+    do_set_realm_authentication_methods,
+    do_set_realm_message_editing,
+    do_set_realm_notifications_stream,
+    do_set_realm_signup_notifications_stream,
+)
 from zerver.actions.streams import (
     bulk_add_subscriptions,
     bulk_remove_subscriptions,
@@ -35,17 +43,7 @@ from zerver.actions.user_settings import (
     do_regenerate_api_key,
 )
 from zerver.actions.users import do_change_user_role, do_deactivate_user
-from zerver.lib.actions import (
-    do_add_realm_domain,
-    do_change_realm_domain,
-    do_deactivate_realm,
-    do_reactivate_realm,
-    do_remove_realm_domain,
-    do_set_realm_authentication_methods,
-    do_set_realm_message_editing,
-    do_set_realm_notifications_stream,
-    do_set_realm_signup_notifications_stream,
-)
+from zerver.lib.actions import do_add_realm_domain, do_change_realm_domain, do_remove_realm_domain
 from zerver.lib.message import get_last_message_id
 from zerver.lib.stream_traffic import get_streams_traffic
 from zerver.lib.streams import create_stream_if_needed
