+++ zerver/tests/test_auth_backends.py
@@ -42,14 +42,14 @@ from social_django.strategy import DjangoStrategy
 from confirmation.models import Confirmation, create_confirmation_link
 from zerver.actions.create_user import do_create_user, do_reactivate_user
 from zerver.actions.invites import do_invite_users
-from zerver.actions.user_settings import do_change_password
-from zerver.actions.users import change_user_is_active, do_deactivate_user
-from zerver.lib.actions import (
-    do_create_realm,
+from zerver.actions.realm_settings import (
     do_deactivate_realm,
     do_reactivate_realm,
     do_set_realm_property,
 )
+from zerver.actions.user_settings import do_change_password
+from zerver.actions.users import change_user_is_active, do_deactivate_user
+from zerver.lib.actions import do_create_realm
 from zerver.lib.avatar import avatar_url
 from zerver.lib.avatar_hash import user_avatar_path
 from zerver.lib.dev_ldap_directory import generate_dev_ldap_dir
