+++ zerver/tests/test_email_mirror.py
@@ -11,9 +11,9 @@ import orjson
 from django.conf import settings
 from django.http import HttpResponse
 
+from zerver.actions.realm_settings import do_deactivate_realm
 from zerver.actions.streams import do_change_stream_post_policy
 from zerver.actions.users import do_deactivate_user
-from zerver.lib.actions import do_deactivate_realm
 from zerver.lib.email_mirror import (
     create_missed_message_address,
     filter_footer,
