+++ zerver/tests/test_event_system.py
@@ -10,8 +10,8 @@ from django.utils.timezone import now as timezone_now
 from version import API_FEATURE_LEVEL, ZULIP_MERGE_BASE, ZULIP_VERSION
 from zerver.actions.message_send import check_send_message
 from zerver.actions.presence import do_update_user_presence
+from zerver.actions.realm_settings import do_set_realm_property
 from zerver.actions.users import do_change_user_role
-from zerver.lib.actions import do_set_realm_property
 from zerver.lib.event_schema import check_restart_event
 from zerver.lib.events import fetch_initial_state_data
 from zerver.lib.exceptions import AccessDeniedError
