+++ zerver/tests/test_retention.py
@@ -7,8 +7,9 @@ from django.utils.timezone import now as timezone_now
 
 from zerver.actions.message_edit import do_delete_messages
 from zerver.actions.message_send import internal_send_private_message
+from zerver.actions.realm_settings import do_set_realm_property
 from zerver.actions.submessage import do_add_submessage
-from zerver.lib.actions import do_create_realm, do_set_realm_property
+from zerver.lib.actions import do_create_realm
 from zerver.lib.retention import (
     archive_messages,
     clean_archived_data,
