+++ zerver/tests/test_signup.py
@@ -36,18 +36,17 @@ from zerver.actions.invites import (
     do_get_invites_controlled_by_user,
     do_invite_users,
 )
+from zerver.actions.realm_settings import (
+    do_deactivate_realm,
+    do_set_realm_property,
+    do_set_realm_user_default_setting,
+)
 from zerver.actions.user_settings import do_change_full_name
 from zerver.actions.users import change_user_is_active, do_change_user_role, do_deactivate_user
 from zerver.context_processors import common_context
 from zerver.decorator import do_two_factor_login
 from zerver.forms import HomepageForm, check_subdomain_available
-from zerver.lib.actions import (
-    do_change_realm_subdomain,
-    do_create_realm,
-    do_deactivate_realm,
-    do_set_realm_property,
-    do_set_realm_user_default_setting,
-)
+from zerver.lib.actions import do_change_realm_subdomain, do_create_realm
 from zerver.lib.email_notifications import enqueue_welcome_emails, followup_day2_email_delay
 from zerver.lib.initial_password import initial_password
 from zerver.lib.mobile_auth_otp import (
