+++ zerver/tests/test_user_groups.py
@@ -5,8 +5,8 @@ from unittest import mock
 import orjson
 from django.utils.timezone import now as timezone_now
 
+from zerver.actions.realm_settings import do_set_realm_property
 from zerver.actions.user_groups import promote_new_full_members
-from zerver.lib.actions import do_set_realm_property
 from zerver.lib.streams import ensure_stream
 from zerver.lib.test_classes import ZulipTestCase
 from zerver.lib.test_helpers import most_recent_usermessage
